Module bookdata::arrow

Re-exports

pub use row_de::RecordBatchDeserializer;

pub use row_de::scan_parquet_file;

pub use writer::TableWriter;

pub use writer::TableWriterBuilder;

pub use types::*;

Modules

batch_io

fusion

Helpers for DataFusion.

row_de

Code to deserialize from Arrow with serde.

types

writer

Traits

FileReader

Parquet file reader API. With this, user can get metadata information about the Parquet file, can get reader for each row group, and access record iterator.

RecordWriter

Functions

open_parquet_file

Convenience function for opening a Parquet reader.

Derive Macros

ParquetRecordWriter

Derive flat, simple RecordWriter implementations. Works by parsing a struct tagged with #[derive(ParquetRecordWriter)] and emitting the correct writing code for each field of the struct. Column writers are generated in the order they are defined.