Module bookdata::arrow::row

Table row interface.

Macros

col_type 🔒

Structs

FrameRecordIter

Iterator implementation for the rows in a data frame.

Enums

RowError

Traits

ColType

Trait for column types.

FrameBuilder

Interface for data frame builders.

FrameStruct

Interface for data frame structs for deserialization.

MappableColType

Marker trait for column types that can be mapped with Into

TableRow

Functions

convert_naive_date 🔒

convert_to_naive_date 🔒

iter_df_rows

Convert a data frame into a vector.

vec_to_df

Convert a vector of records to a chunk.