Trait bookdata::arrow::ArrowSerialize

pub trait ArrowSerialize: ArrowField {
    type MutableArrayType: MutableArray;

    fn new_array() -> Self::MutableArrayType;
    fn arrow_serialize(
        v: &Self::Type,
        array: &mut Self::MutableArrayType
    ) -> Result<(), Error>;
}

Trait that is implemented by all types that are serializable to Arrow.

Implementations are provided for all built-in arrow types as well as Vec, and Option if T implements ArrowSerialize.

Note that Vec implementation needs to be enabled by the crate::arrow_enable_vec_for_type macro.

Required Associated Types

type MutableArrayType: MutableArray

The arrow2::array::MutableArray that holds this value

Required Methods

fn new_array() -> Self::MutableArrayType

Create a new mutable array

fn arrow_serialize(
v: &Self::Type,
array: &mut Self::MutableArrayType
) -> Result<(), Error>

Serialize this field to arrow