Trait bookdata::arrow::FileReader

pub trait FileReader: Send + Sync {
    fn metadata(&self) -> &ParquetMetaData;

    fn num_row_groups(&self) -> usize;

    fn get_row_group(
        &self, 
        i: usize
    ) -> Result<Box<dyn RowGroupReader, Global>, ParquetError>;

    fn get_row_iter(
        &self, 
        projection: Option<Type>
    ) -> Result<RowIter<'_>, ParquetError>;
}

Parquet file reader API. With this, user can get metadata information about the Parquet file, can get reader for each row group, and access record iterator.

Required methods

fn metadata(&self) -> &ParquetMetaData

Get metadata information about this file.

fn num_row_groups(&self) -> usize

Get the total number of row groups for this file.

fn get_row_group(
&self,
i: usize
) -> Result<Box<dyn RowGroupReader, Global>, ParquetError>

Get the ith row group reader. Note this doesn’t do bound check.

fn get_row_iter(
&self,
projection: Option<Type>
) -> Result<RowIter<'_>, ParquetError>

Get full iterator of Rows from a file (over all row groups).

Iterator will automatically load the next row group to advance.

Projected schema can be a subset of or equal to the file schema, when it is None, full file schema is assumed.