bookdata::arrow
Module writer

Structs

TableWriter

Parquet table writer.

Constants

BATCH_SIZE 🔒

ZSTD_LEVEL 🔒

Functions

open_parquet_writer

Open an Arrow Parquet writer using BookData defaults.

open_plpq_writer 🔒

Open a Parquet writer using BookData defaults.

open_polars_writer

Open a Polars data frame Parquet writer using BookData defaults.

parquet_writer_defaults

save_df_parquet

Save a data frame to a Parquet file.

save_df_parquet_nonnull

Save a data frame to a Prquet file without nulls in the schema.