bookdata::cli::filter_marc

Function scan_records

fn scan_records(
    path: &Path,
    filter: &FilterSpec,
    out: impl ObjectWriter<FieldRecord> + Send,
) -> Result<(usize, usize)>

Scan MARC records from a file.

Failes quickly if there is an error opening the file; errors reading the file are from the thread and are availabl when it is joined.