Module bookdata::cli::fusion

DataFusion script runner.

Structs

Fusion

Run a DataFusion script and save its results.

ScriptContext

Functions

async_wrap_errs

Helper function to wrap asynchronous tasks with errors in Molt

cmd_query

Run a query and show its results.

cmd_save_results

Save results from a query.

cmd_table

Bind a table.

save_csv

Save an execution plan results to a CSV file.

save_csvgz

save_parquet

Save an execution plan results to a Parquet directory.

wrap_errs

Helper function to wrap Rust errors in Molt

write_plan_results

Save an execution plan to a writer.