Struct bookdata::cli::goodreads::cluster::ClusterOp

pub struct ClusterOp {
    actions: ActionType,
    data: SrcType,
    clusters: AggType,
    output: PathBuf,
}

Fields
actions: ActionTypedata: SrcTypeclusters: AggTypeoutput: PathBuf
Implementations

impl ClusterOp

pub fn add_actions<P: AsRef<Path>>(path: P) -> ClusterOp

Start a new action-clustering operation.

pub fn ratings<P: AsRef<Path>>(path: P) -> ClusterOp

Start a new rating-clustering operation.

pub fn simple(self) -> ClusterOp

Set operation to cluster simple records instead of full records.

pub fn native_works(self) -> ClusterOp

Set operation to cluster with native works instead of clusters.

pub fn cluster(self) -> Result<()>

Run the clustering operation.

fn load_input(&self) -> PolarsResult<LazyFrame>

Load the input.

fn filter(&self, frame: LazyFrame) -> LazyFrame

Filter the data frame to only the actions we want

fn id_col(&self) -> Expr

Create an identity column reference.

fn project_and_sort(&self, frame: LazyFrame) -> LazyFrame

Project and sort (if possible) the data.

fn aggregates(&self) -> Vec<Expr>

Aggreate the interactions.

fn maybe_integrate_ratings(
&self,
actions: LazyFrame,
source: &LazyFrame
) -> LazyFrame