Struct bookdata::gender::GenderBag

pub struct GenderBag {
    size: usize,
    mask: u32,
    resolved: Gender,
}

A collection of genders.

Fields
size: usizemask: u32resolved: Gender
Implementations

impl GenderBag

pub fn add(&mut self, gender: Gender)

Add a gender to this bag.

pub fn merge_from(&mut self, bag: &GenderBag)

Merge another gender bag into this one.

pub fn len(&self) -> usize

Get the number of gender entries recorded to make this gender record.

pub fn is_empty(&self) -> bool

Check whether the gender bag is empty.

pub fn maybe_gender(&self) -> Option<&Gender>

Get the gender, returning None if no genders are seen.

pub fn to_gender(&self) -> &Gender

Get the gender, returning Gender::Unknown if no genders are seen.