Type Alias bookdata::goodreads::ids::BookId

pub type BookId = i32;

Implementations

impl i32

pub const MIN: i32 = -2_147_483_648i32

The smallest value that can be represented by this integer type (−231).

Examples

Basic usage:

assert_eq!(i32::MIN, -2147483648);

pub const MAX: i32 = 2_147_483_647i32

The largest value that can be represented by this integer type (231 − 1).

Examples

Basic usage:

assert_eq!(i32::MAX, 2147483647);

pub const BITS: u32 = 32u32

The size of this integer type in bits.

Examples

assert_eq!(i32::BITS, 32);

pub fn from_str_radix(src: &str, radix: u32) -> Result<i32, ParseIntError>

Converts a string slice in a given base to an integer.

The string is expected to be an optional + or - sign followed by digits. Leading and trailing whitespace represent an error. Digits are a subset of these characters, depending on radix:

0-9

a-z

A-Z

Panics

This function panics if radix is not in the range from 2 to 36.

Examples

Basic usage:

assert_eq!(i32::from_str_radix("A", 16), Ok(10));

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

Basic usage:

let n = 0b100_0000i32;

assert_eq!(n.count_ones(), 1);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

Basic usage:

assert_eq!(i32::MAX.count_zeros(), 1);

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

Depending on what you’re doing with the value, you might also be interested in the ilog2 function which returns a consistent number, even if the type widens.

Examples

Basic usage:

let n = -1i32;

assert_eq!(n.leading_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

Basic usage:

let n = -4i32;

assert_eq!(n.trailing_zeros(), 2);

pub const fn leading_ones(self) -> u32

Returns the number of leading ones in the binary representation of self.

Examples

Basic usage:

let n = -1i32;

assert_eq!(n.leading_ones(), 32);

pub const fn trailing_ones(self) -> u32

Returns the number of trailing ones in the binary representation of self.

Examples

Basic usage:

let n = 3i32;

assert_eq!(n.trailing_ones(), 2);

pub const fn rotate_left(self, n: u32) -> i32

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isn’t the same operation as the << shifting operator!

Examples

Basic usage:

let n = 0x10000b3i32;
let m = 0xb301;

assert_eq!(n.rotate_left(8), m);

pub const fn rotate_right(self, n: u32) -> i32

Shifts the bits to the right by a specified amount, n, wrapping the truncated bits to the beginning of the resulting integer.

Please note this isn’t the same operation as the >> shifting operator!

Examples

Basic usage:

let n = 0xb301i32;
let m = 0x10000b3;

assert_eq!(n.rotate_right(8), m);

pub const fn swap_bytes(self) -> i32

Reverses the byte order of the integer.

Examples

Basic usage:

let n = 0x12345678i32;

let m = n.swap_bytes();

assert_eq!(m, 0x78563412);

pub const fn reverse_bits(self) -> i32

Reverses the order of bits in the integer. The least significant bit becomes the most significant bit, second least-significant bit becomes second most-significant bit, etc.

Examples

Basic usage:

let n = 0x12345678i32;
let m = n.reverse_bits();

assert_eq!(m, 0x1e6a2c48);
assert_eq!(0, 0i32.reverse_bits());

pub const fn from_be(x: i32) -> i32

Converts an integer from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Ai32;

if cfg!(target_endian = "big") {
    assert_eq!(i32::from_be(n), n)
} else {
    assert_eq!(i32::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: i32) -> i32

Converts an integer from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Ai32;

if cfg!(target_endian = "little") {
    assert_eq!(i32::from_le(n), n)
} else {
    assert_eq!(i32::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> i32

Converts self to big endian from the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Ai32;

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> i32

Converts self to little endian from the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

Basic usage:

let n = 0x1Ai32;

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn checked_add(self, rhs: i32) -> Option<i32>

Checked integer addition. Computes self + rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!((i32::MAX - 2).checked_add(1), Some(i32::MAX - 1));
assert_eq!((i32::MAX - 2).checked_add(3), None);

pub unsafe fn unchecked_add(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked integer addition. Computes self + rhs, assuming overflow cannot occur.

Safety

This results in undefined behavior when self + rhs > i32::MAX or self + rhs < i32::MIN, i.e. when checked_add would return None.

pub const fn checked_add_unsigned(self, rhs: u32) -> Option<i32>

Checked addition with an unsigned integer. Computes self + rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(1i32.checked_add_unsigned(2), Some(3));
assert_eq!((i32::MAX - 2).checked_add_unsigned(3), None);

pub const fn checked_sub(self, rhs: i32) -> Option<i32>

Checked integer subtraction. Computes self - rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!((i32::MIN + 2).checked_sub(1), Some(i32::MIN + 1));
assert_eq!((i32::MIN + 2).checked_sub(3), None);

pub unsafe fn unchecked_sub(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked integer subtraction. Computes self - rhs, assuming overflow cannot occur.

Safety

This results in undefined behavior when self - rhs > i32::MAX or self - rhs < i32::MIN, i.e. when checked_sub would return None.

pub const fn checked_sub_unsigned(self, rhs: u32) -> Option<i32>

Checked subtraction with an unsigned integer. Computes self - rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(1i32.checked_sub_unsigned(2), Some(-1));
assert_eq!((i32::MIN + 2).checked_sub_unsigned(3), None);

pub const fn checked_mul(self, rhs: i32) -> Option<i32>

Checked integer multiplication. Computes self * rhs, returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(i32::MAX.checked_mul(1), Some(i32::MAX));
assert_eq!(i32::MAX.checked_mul(2), None);

pub unsafe fn unchecked_mul(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked integer multiplication. Computes self * rhs, assuming overflow cannot occur.

Safety

This results in undefined behavior when self * rhs > i32::MAX or self * rhs < i32::MIN, i.e. when checked_mul would return None.

pub const fn checked_div(self, rhs: i32) -> Option<i32>

Checked integer division. Computes self / rhs, returning None if rhs == 0 or the division results in overflow.

Examples

Basic usage:

assert_eq!((i32::MIN + 1).checked_div(-1), Some(2147483647));
assert_eq!(i32::MIN.checked_div(-1), None);
assert_eq!((1i32).checked_div(0), None);

pub const fn checked_div_euclid(self, rhs: i32) -> Option<i32>

Checked Euclidean division. Computes self.div_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

Examples

Basic usage:

assert_eq!((i32::MIN + 1).checked_div_euclid(-1), Some(2147483647));
assert_eq!(i32::MIN.checked_div_euclid(-1), None);
assert_eq!((1i32).checked_div_euclid(0), None);

pub const fn checked_rem(self, rhs: i32) -> Option<i32>

Checked integer remainder. Computes self % rhs, returning None if rhs == 0 or the division results in overflow.

Examples

Basic usage:

assert_eq!(5i32.checked_rem(2), Some(1));
assert_eq!(5i32.checked_rem(0), None);
assert_eq!(i32::MIN.checked_rem(-1), None);

pub const fn checked_rem_euclid(self, rhs: i32) -> Option<i32>

Checked Euclidean remainder. Computes self.rem_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

Examples

Basic usage:

assert_eq!(5i32.checked_rem_euclid(2), Some(1));
assert_eq!(5i32.checked_rem_euclid(0), None);
assert_eq!(i32::MIN.checked_rem_euclid(-1), None);

pub const fn checked_neg(self) -> Option<i32>

Checked negation. Computes -self, returning None if self == MIN.

Examples

Basic usage:

assert_eq!(5i32.checked_neg(), Some(-5));
assert_eq!(i32::MIN.checked_neg(), None);

pub const fn checked_shl(self, rhs: u32) -> Option<i32>

Checked shift left. Computes self << rhs, returning None if rhs is larger than or equal to the number of bits in self.

Examples

Basic usage:

assert_eq!(0x1i32.checked_shl(4), Some(0x10));
assert_eq!(0x1i32.checked_shl(129), None);

pub unsafe fn unchecked_shl(self, rhs: u32) -> i32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked shift left. Computes self << rhs, assuming that rhs is less than the number of bits in self.

Safety

This results in undefined behavior if rhs is larger than or equal to the number of bits in self, i.e. when checked_shl would return None.

pub const fn checked_shr(self, rhs: u32) -> Option<i32>

Checked shift right. Computes self >> rhs, returning None if rhs is larger than or equal to the number of bits in self.

Examples

Basic usage:

assert_eq!(0x10i32.checked_shr(4), Some(0x1));
assert_eq!(0x10i32.checked_shr(128), None);

pub unsafe fn unchecked_shr(self, rhs: u32) -> i32

🔬This is a nightly-only experimental API. (unchecked_math)

Unchecked shift right. Computes self >> rhs, assuming that rhs is less than the number of bits in self.

Safety

This results in undefined behavior if rhs is larger than or equal to the number of bits in self, i.e. when checked_shr would return None.

pub const fn checked_abs(self) -> Option<i32>

Checked absolute value. Computes self.abs(), returning None if self == MIN.

Examples

Basic usage:

assert_eq!((-5i32).checked_abs(), Some(5));
assert_eq!(i32::MIN.checked_abs(), None);

pub const fn checked_pow(self, exp: u32) -> Option<i32>

Checked exponentiation. Computes self.pow(exp), returning None if overflow occurred.

Examples

Basic usage:

assert_eq!(8i32.checked_pow(2), Some(64));
assert_eq!(i32::MAX.checked_pow(2), None);

pub fn checked_isqrt(self) -> Option<i32>

🔬This is a nightly-only experimental API. (isqrt)

Returns the square root of the number, rounded down.

Returns None if self is negative.

Examples

Basic usage:

#![feature(isqrt)]
assert_eq!(10i32.checked_isqrt(), Some(3));

pub const fn saturating_add(self, rhs: i32) -> i32

Saturating integer addition. Computes self + rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(100i32.saturating_add(1), 101);
assert_eq!(i32::MAX.saturating_add(100), i32::MAX);
assert_eq!(i32::MIN.saturating_add(-1), i32::MIN);

pub const fn saturating_add_unsigned(self, rhs: u32) -> i32

Saturating addition with an unsigned integer. Computes self + rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(1i32.saturating_add_unsigned(2), 3);
assert_eq!(i32::MAX.saturating_add_unsigned(100), i32::MAX);

pub const fn saturating_sub(self, rhs: i32) -> i32

Saturating integer subtraction. Computes self - rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(100i32.saturating_sub(127), -27);
assert_eq!(i32::MIN.saturating_sub(100), i32::MIN);
assert_eq!(i32::MAX.saturating_sub(-1), i32::MAX);

pub const fn saturating_sub_unsigned(self, rhs: u32) -> i32

Saturating subtraction with an unsigned integer. Computes self - rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(100i32.saturating_sub_unsigned(127), -27);
assert_eq!(i32::MIN.saturating_sub_unsigned(100), i32::MIN);

pub const fn saturating_neg(self) -> i32

Saturating integer negation. Computes -self, returning MAX if self == MIN instead of overflowing.

Examples

Basic usage:

assert_eq!(100i32.saturating_neg(), -100);
assert_eq!((-100i32).saturating_neg(), 100);
assert_eq!(i32::MIN.saturating_neg(), i32::MAX);
assert_eq!(i32::MAX.saturating_neg(), i32::MIN + 1);

pub const fn saturating_abs(self) -> i32

Saturating absolute value. Computes self.abs(), returning MAX if self == MIN instead of overflowing.

Examples

Basic usage:

assert_eq!(100i32.saturating_abs(), 100);
assert_eq!((-100i32).saturating_abs(), 100);
assert_eq!(i32::MIN.saturating_abs(), i32::MAX);
assert_eq!((i32::MIN + 1).saturating_abs(), i32::MAX);

pub const fn saturating_mul(self, rhs: i32) -> i32

Saturating integer multiplication. Computes self * rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(10i32.saturating_mul(12), 120);
assert_eq!(i32::MAX.saturating_mul(10), i32::MAX);
assert_eq!(i32::MIN.saturating_mul(10), i32::MIN);

pub const fn saturating_div(self, rhs: i32) -> i32

Saturating integer division. Computes self / rhs, saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!(5i32.saturating_div(2), 2);
assert_eq!(i32::MAX.saturating_div(-1), i32::MIN + 1);
assert_eq!(i32::MIN.saturating_div(-1), i32::MAX);

let _ = 1i32.saturating_div(0);

pub const fn saturating_pow(self, exp: u32) -> i32

Saturating integer exponentiation. Computes self.pow(exp), saturating at the numeric bounds instead of overflowing.

Examples

Basic usage:

assert_eq!((-4i32).saturating_pow(3), -64);
assert_eq!(i32::MIN.saturating_pow(2), i32::MAX);
assert_eq!(i32::MIN.saturating_pow(3), i32::MIN);

pub const fn wrapping_add(self, rhs: i32) -> i32

Wrapping (modular) addition. Computes self + rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(100i32.wrapping_add(27), 127);
assert_eq!(i32::MAX.wrapping_add(2), i32::MIN + 1);

pub const fn wrapping_add_unsigned(self, rhs: u32) -> i32

Wrapping (modular) addition with an unsigned integer. Computes self + rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(100i32.wrapping_add_unsigned(27), 127);
assert_eq!(i32::MAX.wrapping_add_unsigned(2), i32::MIN + 1);

pub const fn wrapping_sub(self, rhs: i32) -> i32

Wrapping (modular) subtraction. Computes self - rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(0i32.wrapping_sub(127), -127);
assert_eq!((-2i32).wrapping_sub(i32::MAX), i32::MAX);

pub const fn wrapping_sub_unsigned(self, rhs: u32) -> i32

Wrapping (modular) subtraction with an unsigned integer. Computes self - rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(0i32.wrapping_sub_unsigned(127), -127);
assert_eq!((-2i32).wrapping_sub_unsigned(u32::MAX), -1);

pub const fn wrapping_mul(self, rhs: i32) -> i32

Wrapping (modular) multiplication. Computes self * rhs, wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(10i32.wrapping_mul(12), 120);
assert_eq!(11i8.wrapping_mul(12), -124);

pub const fn wrapping_div(self, rhs: i32) -> i32

Wrapping (modular) division. Computes self / rhs, wrapping around at the boundary of the type.

The only case where such wrapping can occur is when one divides MIN / -1 on a signed type (where MIN is the negative minimal value for the type); this is equivalent to -MIN, a positive value that is too large to represent in the type. In such a case, this function returns MIN itself.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100i32.wrapping_div(10), 10);
assert_eq!((-128i8).wrapping_div(-1), -128);

pub const fn wrapping_div_euclid(self, rhs: i32) -> i32

Wrapping Euclidean division. Computes self.div_euclid(rhs), wrapping around at the boundary of the type.

Wrapping will only occur in MIN / -1 on a signed type (where MIN is the negative minimal value for the type). This is equivalent to -MIN, a positive value that is too large to represent in the type. In this case, this method returns MIN itself.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100i32.wrapping_div_euclid(10), 10);
assert_eq!((-128i8).wrapping_div_euclid(-1), -128);

pub const fn wrapping_rem(self, rhs: i32) -> i32

Wrapping (modular) remainder. Computes self % rhs, wrapping around at the boundary of the type.

Such wrap-around never actually occurs mathematically; implementation artifacts make x % y invalid for MIN / -1 on a signed type (where MIN is the negative minimal value). In such a case, this function returns 0.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100i32.wrapping_rem(10), 0);
assert_eq!((-128i8).wrapping_rem(-1), 0);

pub const fn wrapping_rem_euclid(self, rhs: i32) -> i32

Wrapping Euclidean remainder. Computes self.rem_euclid(rhs), wrapping around at the boundary of the type.

Wrapping will only occur in MIN % -1 on a signed type (where MIN is the negative minimal value for the type). In this case, this method returns 0.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(100i32.wrapping_rem_euclid(10), 0);
assert_eq!((-128i8).wrapping_rem_euclid(-1), 0);

pub const fn wrapping_neg(self) -> i32

Wrapping (modular) negation. Computes -self, wrapping around at the boundary of the type.

The only case where such wrapping can occur is when one negates MIN on a signed type (where MIN is the negative minimal value for the type); this is a positive value that is too large to represent in the type. In such a case, this function returns MIN itself.

Examples

Basic usage:

assert_eq!(100i32.wrapping_neg(), -100);
assert_eq!(i32::MIN.wrapping_neg(), i32::MIN);

pub const fn wrapping_shl(self, rhs: u32) -> i32

Panic-free bitwise shift-left; yields self << mask(rhs), where mask removes any high-order bits of rhs that would cause the shift to exceed the bitwidth of the type.

Note that this is not the same as a rotate-left; the RHS of a wrapping shift-left is restricted to the range of the type, rather than the bits shifted out of the LHS being returned to the other end. The primitive integer types all implement a rotate_left function, which may be what you want instead.

Examples

Basic usage:

assert_eq!((-1i32).wrapping_shl(7), -128);
assert_eq!((-1i32).wrapping_shl(128), -1);

pub const fn wrapping_shr(self, rhs: u32) -> i32

Panic-free bitwise shift-right; yields self >> mask(rhs), where mask removes any high-order bits of rhs that would cause the shift to exceed the bitwidth of the type.

Note that this is not the same as a rotate-right; the RHS of a wrapping shift-right is restricted to the range of the type, rather than the bits shifted out of the LHS being returned to the other end. The primitive integer types all implement a rotate_right function, which may be what you want instead.

Examples

Basic usage:

assert_eq!((-128i32).wrapping_shr(7), -1);
assert_eq!((-128i16).wrapping_shr(64), -128);

pub const fn wrapping_abs(self) -> i32

Wrapping (modular) absolute value. Computes self.abs(), wrapping around at the boundary of the type.

The only case where such wrapping can occur is when one takes the absolute value of the negative minimal value for the type; this is a positive value that is too large to represent in the type. In such a case, this function returns MIN itself.

Examples

Basic usage:

assert_eq!(100i32.wrapping_abs(), 100);
assert_eq!((-100i32).wrapping_abs(), 100);
assert_eq!(i32::MIN.wrapping_abs(), i32::MIN);
assert_eq!((-128i8).wrapping_abs() as u8, 128);

pub const fn unsigned_abs(self) -> u32

Computes the absolute value of self without any wrapping or panicking.

Examples

Basic usage:

assert_eq!(100i32.unsigned_abs(), 100u32);
assert_eq!((-100i32).unsigned_abs(), 100u32);
assert_eq!((-128i8).unsigned_abs(), 128u8);

pub const fn wrapping_pow(self, exp: u32) -> i32

Wrapping (modular) exponentiation. Computes self.pow(exp), wrapping around at the boundary of the type.

Examples

Basic usage:

assert_eq!(3i32.wrapping_pow(4), 81);
assert_eq!(3i8.wrapping_pow(5), -13);
assert_eq!(3i8.wrapping_pow(6), -39);

pub const fn overflowing_add(self, rhs: i32) -> (i32, bool)

Calculates self + rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

assert_eq!(5i32.overflowing_add(2), (7, false));
assert_eq!(i32::MAX.overflowing_add(1), (i32::MIN, true));

pub fn carrying_add(self, rhs: i32, carry: bool) -> (i32, bool)

🔬This is a nightly-only experimental API. (bigint_helper_methods)

Calculates self + rhs + carry and checks for overflow.

Performs “ternary addition” of two integer operands and a carry-in bit, and returns a tuple of the sum along with a boolean indicating whether an arithmetic overflow would occur. On overflow, the wrapped value is returned.

This allows chaining together multiple additions to create a wider addition, and can be useful for bignum addition. This method should only be used for the most significant word; for the less significant words the unsigned method u32::carrying_add should be used.

The output boolean returned by this method is not a carry flag, and should not be added to a more significant word.

If the input carry is false, this method is equivalent to overflowing_add.

Examples

#![feature(bigint_helper_methods)]
// Only the most significant word is signed.
//
//   10  MAX    (a = 10 × 2^32 + 2^32 - 1)
// + -5    9    (b = -5 × 2^32 + 9)
// ---------
//    6    8    (sum = 6 × 2^32 + 8)

let (a1, a0): (i32, u32) = (10, u32::MAX);
let (b1, b0): (i32, u32) = (-5, 9);
let carry0 = false;

// u32::carrying_add for the less significant words
let (sum0, carry1) = a0.carrying_add(b0, carry0);
assert_eq!(carry1, true);

// i32::carrying_add for the most significant word
let (sum1, overflow) = a1.carrying_add(b1, carry1);
assert_eq!(overflow, false);

assert_eq!((sum1, sum0), (6, 8));

pub const fn overflowing_add_unsigned(self, rhs: u32) -> (i32, bool)

Calculates self + rhs with an unsigned rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

assert_eq!(1i32.overflowing_add_unsigned(2), (3, false));
assert_eq!((i32::MIN).overflowing_add_unsigned(u32::MAX), (i32::MAX, false));
assert_eq!((i32::MAX - 2).overflowing_add_unsigned(3), (i32::MIN, true));

pub const fn overflowing_sub(self, rhs: i32) -> (i32, bool)

Calculates self - rhs

Returns a tuple of the subtraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

assert_eq!(5i32.overflowing_sub(2), (3, false));
assert_eq!(i32::MIN.overflowing_sub(1), (i32::MAX, true));

pub fn borrowing_sub(self, rhs: i32, borrow: bool) -> (i32, bool)

🔬This is a nightly-only experimental API. (bigint_helper_methods)

Calculates self − rhs − borrow and checks for overflow.

Performs “ternary subtraction” by subtracting both an integer operand and a borrow-in bit from self, and returns a tuple of the difference along with a boolean indicating whether an arithmetic overflow would occur. On overflow, the wrapped value is returned.

This allows chaining together multiple subtractions to create a wider subtraction, and can be useful for bignum subtraction. This method should only be used for the most significant word; for the less significant words the unsigned method u32::borrowing_sub should be used.

The output boolean returned by this method is not a borrow flag, and should not be subtracted from a more significant word.

If the input borrow is false, this method is equivalent to overflowing_sub.

Examples

#![feature(bigint_helper_methods)]
// Only the most significant word is signed.
//
//    6    8    (a = 6 × 2^32 + 8)
// - -5    9    (b = -5 × 2^32 + 9)
// ---------
//   10  MAX    (diff = 10 × 2^32 + 2^32 - 1)

let (a1, a0): (i32, u32) = (6, 8);
let (b1, b0): (i32, u32) = (-5, 9);
let borrow0 = false;

// u32::borrowing_sub for the less significant words
let (diff0, borrow1) = a0.borrowing_sub(b0, borrow0);
assert_eq!(borrow1, true);

// i32::borrowing_sub for the most significant word
let (diff1, overflow) = a1.borrowing_sub(b1, borrow1);
assert_eq!(overflow, false);

assert_eq!((diff1, diff0), (10, u32::MAX));

pub const fn overflowing_sub_unsigned(self, rhs: u32) -> (i32, bool)

Calculates self - rhs with an unsigned rhs

Returns a tuple of the subtraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

assert_eq!(1i32.overflowing_sub_unsigned(2), (-1, false));
assert_eq!((i32::MAX).overflowing_sub_unsigned(u32::MAX), (i32::MIN, false));
assert_eq!((i32::MIN + 2).overflowing_sub_unsigned(3), (i32::MAX, true));

pub const fn overflowing_mul(self, rhs: i32) -> (i32, bool)

Calculates the multiplication of self and rhs.

Returns a tuple of the multiplication along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Examples

Basic usage:

assert_eq!(5i32.overflowing_mul(2), (10, false));
assert_eq!(1_000_000_000i32.overflowing_mul(10), (1410065408, true));

pub const fn overflowing_div(self, rhs: i32) -> (i32, bool)

Calculates the divisor when self is divided by rhs.

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then self is returned.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(5i32.overflowing_div(2), (2, false));
assert_eq!(i32::MIN.overflowing_div(-1), (i32::MIN, true));

pub const fn overflowing_div_euclid(self, rhs: i32) -> (i32, bool)

Calculates the quotient of Euclidean division self.div_euclid(rhs).

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then self is returned.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(5i32.overflowing_div_euclid(2), (2, false));
assert_eq!(i32::MIN.overflowing_div_euclid(-1), (i32::MIN, true));

pub const fn overflowing_rem(self, rhs: i32) -> (i32, bool)

Calculates the remainder when self is divided by rhs.

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(5i32.overflowing_rem(2), (1, false));
assert_eq!(i32::MIN.overflowing_rem(-1), (0, true));

pub const fn overflowing_rem_euclid(self, rhs: i32) -> (i32, bool)

Overflowing Euclidean remainder. Calculates self.rem_euclid(rhs).

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned.

Panics

This function will panic if rhs is 0.

Examples

Basic usage:

assert_eq!(5i32.overflowing_rem_euclid(2), (1, false));
assert_eq!(i32::MIN.overflowing_rem_euclid(-1), (0, true));

pub const fn overflowing_neg(self) -> (i32, bool)

Negates self, overflowing if this is equal to the minimum value.

Returns a tuple of the negated version of self along with a boolean indicating whether an overflow happened. If self is the minimum value (e.g., i32::MIN for values of type i32), then the minimum value will be returned again and true will be returned for an overflow happening.

Examples

Basic usage:

assert_eq!(2i32.overflowing_neg(), (-2, false));
assert_eq!(i32::MIN.overflowing_neg(), (i32::MIN, true));

pub const fn overflowing_shl(self, rhs: u32) -> (i32, bool)

Shifts self left by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits. If the shift value is too large, then value is masked (N-1) where N is the number of bits, and this value is then used to perform the shift.

Examples

Basic usage:

assert_eq!(0x1i32.overflowing_shl(4), (0x10, false));
assert_eq!(0x1i32.overflowing_shl(36), (0x10, true));

pub const fn overflowing_shr(self, rhs: u32) -> (i32, bool)

Shifts self right by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits. If the shift value is too large, then value is masked (N-1) where N is the number of bits, and this value is then used to perform the shift.

Examples

Basic usage:

assert_eq!(0x10i32.overflowing_shr(4), (0x1, false));
assert_eq!(0x10i32.overflowing_shr(36), (0x1, true));

pub const fn overflowing_abs(self) -> (i32, bool)

Computes the absolute value of self.

Returns a tuple of the absolute version of self along with a boolean indicating whether an overflow happened. If self is the minimum value (e.g., i32::MIN for values of type i32), then the minimum value will be returned again and true will be returned for an overflow happening.

Examples

Basic usage:

assert_eq!(10i32.overflowing_abs(), (10, false));
assert_eq!((-10i32).overflowing_abs(), (10, false));
assert_eq!((i32::MIN).overflowing_abs(), (i32::MIN, true));

pub const fn overflowing_pow(self, exp: u32) -> (i32, bool)

Raises self to the power of exp, using exponentiation by squaring.

Returns a tuple of the exponentiation along with a bool indicating whether an overflow happened.

Examples

Basic usage:

assert_eq!(3i32.overflowing_pow(4), (81, false));
assert_eq!(3i8.overflowing_pow(5), (-13, true));

pub const fn pow(self, exp: u32) -> i32

Raises self to the power of exp, using exponentiation by squaring.

Examples

Basic usage:

let x: i32 = 2; // or any other integer type

assert_eq!(x.pow(5), 32);

pub fn isqrt(self) -> i32

🔬This is a nightly-only experimental API. (isqrt)

Returns the square root of the number, rounded down.

Panics

This function will panic if self is negative.

Examples

Basic usage:

#![feature(isqrt)]
assert_eq!(10i32.isqrt(), 3);

pub const fn div_euclid(self, rhs: i32) -> i32

Calculates the quotient of Euclidean division of self by rhs.

This computes the integer q such that self = q * rhs + r, with r = self.rem_euclid(rhs) and 0 <= r < abs(rhs).

In other words, the result is self / rhs rounded to the integer q such that self >= q * rhs. If self > 0, this is equal to round towards zero (the default in Rust); if self < 0, this is equal to round towards +/- infinity.

Panics

This function will panic if rhs is 0 or the division results in overflow.

Examples

Basic usage:

let a: i32 = 7; // or any other integer type
let b = 4;

assert_eq!(a.div_euclid(b), 1); // 7 >= 4 * 1
assert_eq!(a.div_euclid(-b), -1); // 7 >= -4 * -1
assert_eq!((-a).div_euclid(b), -2); // -7 >= 4 * -2
assert_eq!((-a).div_euclid(-b), 2); // -7 >= -4 * 2

pub const fn rem_euclid(self, rhs: i32) -> i32

Calculates the least nonnegative remainder of self (mod rhs).

This is done as if by the Euclidean division algorithm – given r = self.rem_euclid(rhs), self = rhs * self.div_euclid(rhs) + r, and 0 <= r < abs(rhs).

Panics

This function will panic if rhs is 0 or the division results in overflow.

Examples

Basic usage:

let a: i32 = 7; // or any other integer type
let b = 4;

assert_eq!(a.rem_euclid(b), 3);
assert_eq!((-a).rem_euclid(b), 1);
assert_eq!(a.rem_euclid(-b), 3);
assert_eq!((-a).rem_euclid(-b), 1);

pub const fn div_floor(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (int_roundings)

Calculates the quotient of self and rhs, rounding the result towards negative infinity.

Panics

This function will panic if rhs is zero.

Overflow behavior

On overflow, this function will panic if overflow checks are enabled (default in debug mode) and wrap if overflow checks are disabled (default in release mode).

Examples

Basic usage:

#![feature(int_roundings)]
let a: i32 = 8;
let b = 3;

assert_eq!(a.div_floor(b), 2);
assert_eq!(a.div_floor(-b), -3);
assert_eq!((-a).div_floor(b), -3);
assert_eq!((-a).div_floor(-b), 2);

pub const fn div_ceil(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (int_roundings)

Calculates the quotient of self and rhs, rounding the result towards positive infinity.

Panics

This function will panic if rhs is zero.

Overflow behavior

On overflow, this function will panic if overflow checks are enabled (default in debug mode) and wrap if overflow checks are disabled (default in release mode).

Examples

Basic usage:

#![feature(int_roundings)]
let a: i32 = 8;
let b = 3;

assert_eq!(a.div_ceil(b), 3);
assert_eq!(a.div_ceil(-b), -2);
assert_eq!((-a).div_ceil(b), -2);
assert_eq!((-a).div_ceil(-b), 3);

pub const fn next_multiple_of(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (int_roundings)

If rhs is positive, calculates the smallest value greater than or equal to self that is a multiple of rhs. If rhs is negative, calculates the largest value less than or equal to self that is a multiple of rhs.

Panics

This function will panic if rhs is zero.

Overflow behavior

On overflow, this function will panic if overflow checks are enabled (default in debug mode) and wrap if overflow checks are disabled (default in release mode).

Examples

Basic usage:

#![feature(int_roundings)]
assert_eq!(16_i32.next_multiple_of(8), 16);
assert_eq!(23_i32.next_multiple_of(8), 24);
assert_eq!(16_i32.next_multiple_of(-8), 16);
assert_eq!(23_i32.next_multiple_of(-8), 16);
assert_eq!((-16_i32).next_multiple_of(8), -16);
assert_eq!((-23_i32).next_multiple_of(8), -16);
assert_eq!((-16_i32).next_multiple_of(-8), -16);
assert_eq!((-23_i32).next_multiple_of(-8), -24);

pub const fn checked_next_multiple_of(self, rhs: i32) -> Option<i32>

🔬This is a nightly-only experimental API. (int_roundings)

If rhs is positive, calculates the smallest value greater than or equal to self that is a multiple of rhs. If rhs is negative, calculates the largest value less than or equal to self that is a multiple of rhs. Returns None if rhs is zero or the operation would result in overflow.

Examples

Basic usage:

#![feature(int_roundings)]
assert_eq!(16_i32.checked_next_multiple_of(8), Some(16));
assert_eq!(23_i32.checked_next_multiple_of(8), Some(24));
assert_eq!(16_i32.checked_next_multiple_of(-8), Some(16));
assert_eq!(23_i32.checked_next_multiple_of(-8), Some(16));
assert_eq!((-16_i32).checked_next_multiple_of(8), Some(-16));
assert_eq!((-23_i32).checked_next_multiple_of(8), Some(-16));
assert_eq!((-16_i32).checked_next_multiple_of(-8), Some(-16));
assert_eq!((-23_i32).checked_next_multiple_of(-8), Some(-24));
assert_eq!(1_i32.checked_next_multiple_of(0), None);
assert_eq!(i32::MAX.checked_next_multiple_of(2), None);

pub fn midpoint(self, rhs: i32) -> i32

🔬This is a nightly-only experimental API. (num_midpoint)

Calculates the middle point of self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

#![feature(num_midpoint)]
assert_eq!(0i32.midpoint(4), 2);
assert_eq!(0i32.midpoint(-1), -1);
assert_eq!((-1i32).midpoint(0), -1);

pub const fn ilog(self, base: i32) -> u32

Returns the logarithm of the number with respect to an arbitrary base, rounded down.

This method might not be optimized owing to implementation details; ilog2 can produce results more efficiently for base 2, and ilog10 can produce results more efficiently for base 10.

Panics

This function will panic if self is less than or equal to zero, or if base is less than 2.

Examples

assert_eq!(5i32.ilog(5), 1);

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

Panics

This function will panic if self is less than or equal to zero.

Examples

assert_eq!(2i32.ilog2(), 1);

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

Panics

This function will panic if self is less than or equal to zero.

Example

assert_eq!(10i32.ilog10(), 1);

pub const fn checked_ilog(self, base: i32) -> Option<u32>

Returns the logarithm of the number with respect to an arbitrary base, rounded down.

Returns None if the number is negative or zero, or if the base is not at least 2.

This method might not be optimized owing to implementation details; checked_ilog2 can produce results more efficiently for base 2, and checked_ilog10 can produce results more efficiently for base 10.

Examples

assert_eq!(5i32.checked_ilog(5), Some(1));

pub const fn checked_ilog2(self) -> Option<u32>

Returns the base 2 logarithm of the number, rounded down.

Returns None if the number is negative or zero.

Examples

assert_eq!(2i32.checked_ilog2(), Some(1));

pub const fn checked_ilog10(self) -> Option<u32>

Returns the base 10 logarithm of the number, rounded down.

Returns None if the number is negative or zero.

Example

assert_eq!(10i32.checked_ilog10(), Some(1));

pub const fn abs(self) -> i32

Computes the absolute value of self.

Overflow behavior

The absolute value of i32::MIN cannot be represented as an i32, and attempting to calculate it will cause an overflow. This means that code in debug mode will trigger a panic on this case and optimized code will return i32::MIN without a panic.

Examples

Basic usage:

assert_eq!(10i32.abs(), 10);
assert_eq!((-10i32).abs(), 10);

pub const fn abs_diff(self, other: i32) -> u32

Computes the absolute difference between self and other.

This function always returns the correct answer without overflow or panics by returning an unsigned integer.

Examples

Basic usage:

assert_eq!(100i32.abs_diff(80), 20u32);
assert_eq!(100i32.abs_diff(110), 10u32);
assert_eq!((-100i32).abs_diff(80), 180u32);
assert_eq!((-100i32).abs_diff(-120), 20u32);
assert_eq!(i32::MIN.abs_diff(i32::MAX), u32::MAX);

pub const fn signum(self) -> i32

Returns a number representing sign of self.

0 if the number is zero

1 if the number is positive

-1 if the number is negative

Examples

Basic usage:

assert_eq!(10i32.signum(), 1);
assert_eq!(0i32.signum(), 0);
assert_eq!((-10i32).signum(), -1);

pub const fn is_positive(self) -> bool

Returns true if self is positive and false if the number is zero or negative.

Examples

Basic usage:

assert!(10i32.is_positive());
assert!(!(-10i32).is_positive());

pub const fn is_negative(self) -> bool

Returns true if self is negative and false if the number is zero or positive.

Examples

Basic usage:

assert!((-10i32).is_negative());
assert!(!10i32.is_negative());

pub const fn to_be_bytes(self) -> [u8; 4]

Return the memory representation of this integer as a byte array in big-endian (network) byte order.

Examples

let bytes = 0x12345678i32.to_be_bytes();
assert_eq!(bytes, [0x12, 0x34, 0x56, 0x78]);

pub const fn to_le_bytes(self) -> [u8; 4]

Return the memory representation of this integer as a byte array in little-endian byte order.

Examples

let bytes = 0x12345678i32.to_le_bytes();
assert_eq!(bytes, [0x78, 0x56, 0x34, 0x12]);

pub const fn to_ne_bytes(self) -> [u8; 4]

Return the memory representation of this integer as a byte array in native byte order.

As the target platform’s native endianness is used, portable code should use to_be_bytes or to_le_bytes, as appropriate, instead.

Examples

let bytes = 0x12345678i32.to_ne_bytes();
assert_eq!(
    bytes,
    if cfg!(target_endian = "big") {
        [0x12, 0x34, 0x56, 0x78]
    } else {
        [0x78, 0x56, 0x34, 0x12]
    }
);

pub const fn from_be_bytes(bytes: [u8; 4]) -> i32

Create an integer value from its representation as a byte array in big endian.

Examples

let value = i32::from_be_bytes([0x12, 0x34, 0x56, 0x78]);
assert_eq!(value, 0x12345678);

When starting from a slice rather than an array, fallible conversion APIs can be used:

fn read_be_i32(input: &mut &[u8]) -> i32 {
    let (int_bytes, rest) = input.split_at(std::mem::size_of::<i32>());
    *input = rest;
    i32::from_be_bytes(int_bytes.try_into().unwrap())
}

pub const fn from_le_bytes(bytes: [u8; 4]) -> i32

Create an integer value from its representation as a byte array in little endian.

Examples

let value = i32::from_le_bytes([0x78, 0x56, 0x34, 0x12]);
assert_eq!(value, 0x12345678);

When starting from a slice rather than an array, fallible conversion APIs can be used:

fn read_le_i32(input: &mut &[u8]) -> i32 {
    let (int_bytes, rest) = input.split_at(std::mem::size_of::<i32>());
    *input = rest;
    i32::from_le_bytes(int_bytes.try_into().unwrap())
}

pub const fn from_ne_bytes(bytes: [u8; 4]) -> i32

Create an integer value from its memory representation as a byte array in native endianness.

As the target platform’s native endianness is used, portable code likely wants to use from_be_bytes or from_le_bytes, as appropriate instead.

Examples

let value = i32::from_ne_bytes(if cfg!(target_endian = "big") {
    [0x12, 0x34, 0x56, 0x78]
} else {
    [0x78, 0x56, 0x34, 0x12]
});
assert_eq!(value, 0x12345678);

When starting from a slice rather than an array, fallible conversion APIs can be used:

fn read_ne_i32(input: &mut &[u8]) -> i32 {
    let (int_bytes, rest) = input.split_at(std::mem::size_of::<i32>());
    *input = rest;
    i32::from_ne_bytes(int_bytes.try_into().unwrap())
}

pub const fn min_value() -> i32

👎Deprecating in a future Rust version: replaced by the MIN associated constant on this type

New code should prefer to use i32::MIN instead.

Returns the smallest value that can be represented by this integer type.

pub const fn max_value() -> i32

👎Deprecating in a future Rust version: replaced by the MAX associated constant on this type

New code should prefer to use i32::MAX instead.

Returns the largest value that can be represented by this integer type.