Type Alias bookdata::goodreads::users::UserIndex

pub type UserIndex = IdIndex<String>;

Aliased Type

struct UserIndex {
    map: HashMap<String, i32, BuildHasherDefault<AHasher>, Global>,
    frozen: bool,
}

Fields
map: HashMap<String, i32, BuildHasherDefault<AHasher>, Global>frozen: bool
Implementations

impl<K> IdIndex<K>where K: Eq + Hash,

pub fn new() -> IdIndex<K>

Create a new index.

pub fn freeze(self) -> IdIndex<K>

Freeze the index so no new items can be added.

pub fn len(&self) -> usize

Get the index length

pub fn intern<Q>(&mut self, key: &Q) -> Result<i32, IndexError>where K: Borrow<Q>, Q: Hash + Eq + ToOwned<Owned = K> + ?Sized,

Get the ID for a key, adding it to the index if needed.

pub fn intern_owned(&mut self, key: K) -> Result<i32, IndexError>

Get the ID for a key, adding it to the index if needed and transferring ownership.

pub fn lookup<Q>(&self, key: &Q) -> Option<i32>where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Look up the ID for a key if it is present.

pub fn keys(&self) -> Keys<'_, K, i32>

Iterate over keys (see std::collections::HashMap::keys).

impl IdIndex<String>

pub fn key_vec(&self) -> Vec<&str>

Get the keys in order.

pub fn data_frame( &self, id_col: &str, key_col: &str ) -> Result<DataFrame, PolarsError>

Conver this ID index into a DataFrame, with columns for ID and key.

pub fn load_standard<P: AsRef<Path>>(path: P) -> Result<IdIndex<String>>

Load from a Parquet file, with a standard configuration.

This assumes the Parquet file has the following columns:

key, of type String, storing the keys

id, of type i32, storing the IDs

pub fn load<P: AsRef<Path>>( path: P, id_col: &str, key_col: &str ) -> Result<IdIndex<String>>

Load from a Parquet file.

This loads two columns from a Parquet file. The ID column is expected to have type UInt32 (or a type projectable to it), and the key column should be Utf8.

pub fn load_csv<P: AsRef<Path>, K: Eq + Hash + DeserializeOwned>( path: P ) -> Result<IdIndex<K>>

Load an index from a CSV file.

This loads an index from a CSV file. It assumes the first column is the ID, and the second column is the key.

pub fn save_standard<P: AsRef<Path>>(&self, path: P) -> Result<()>

Save to a Parquet file with the standard configuration.

pub fn save<P: AsRef<Path>>( &self, path: P, id_col: &str, key_col: &str ) -> Result<()>

Save to a Parquet file with the standard configuration.