bookdata::ids::codes

Struct NS

pub struct NS<'a> {
    pub name: &'a str,
    pub fn_name: &'a str,
    pub code: i32,
}

The “number space” structure for identifier spaces.

Fields
name: &'a str

The name of this numberspace.

fn_name: &'a str

The name usable as a part of an identifier name (snake_case).

code: i32

The number space numeric code.

Implementations

impl<'a> NS<'a>

const fn new(name: &'a str, fn_name: &'a str, code: i32) -> NS<'a>

Create a new number space. Internal only.

pub fn name(&'a self) -> &'a str

Get the name of the number space.

pub fn code(&'a self) -> i32

Get the numeric code of the number space.

pub fn base(&'a self) -> i32

Get the base of the number space. Identifiers are translated into this space by adding the base.

pub fn to_code(&'a self, n: i32) -> i32

Convert a numeric identifier to a book code in this number space.

pub fn from_code(&'a self, n: i32) -> Option<i32>

Extract a numeric identifier from a book code in this number space.

impl NS<'static>

pub fn by_name(name: &str) -> Option<&'static NS<'static>>