Module bookdata::interactions

Manage and deduplicate ratings and interactions.

This code consolidates rating de-duplication into a single place, so we can use the same logic across data sets. We always store timestamps, dropping them at output time, because the largest data sets have timestamps. Saving space for the smallest data set doesn’t seem worthwhile.

Re-exports

pub use actions::*;

pub use ratings::*;

Modules

actions 🔒

ratings 🔒

Structs

Key 🔒

Traits

Dedup

Interface for de-duplicating interactions.

Interaction

Trait for an interaction.