bookdata::io::lines

Struct LineProcessor

pub struct LineProcessor {
    reader: Box<dyn BufRead>,
}

Read lines from a file with buffering, decompression, and parsing.

Fields
reader: Box<dyn BufRead>
Implementations

impl LineProcessor

pub fn open_gzip(path: &Path, pb: ProgressBar) -> Result<LineProcessor>

Open a line processor from a gzipped source.

pub fn lines(self) -> Lines<Box<dyn BufRead>>

Get the lines as strings.

pub fn records<R: FromStr>(self) -> Records<R>

Get the lines as records.

pub fn json_records<R: DeserializeOwned>(self) -> JSONRecords<R>

Get the lines as JSON records.

pub fn process_json<W, R>(self, writer: &mut W) -> Result<usize>
where R: DeserializeOwned, W: ObjectWriter<R>,

Process JSON rows into an object writer.

This parses each line of the data set, deserializes it with JSON, and passes the resulting object to the specified ObjectWriter. It produces useful error messages with line numbers when there is a failure. It does not call ObjectWriter::finish when it is done - the caller needs to do that.