bookdata::marc

Module record

MARC record representation.

The code in this module supports records in the MARC format. It can be used for processing both bibliographic and name authority records.

Structs

Code

An indicator or subfield code.

ControlField

A control field (00X) in a MARC record.

Field

A field in a MARC record.

MARCRecord

A MARC record.

Subfield

A subfield in a MARC record.

Enums

ParseCodeError

Error parsing a code