Module bookdata::parsing

Parsing utilities. Format-specific parsing is in various other modules.

Re-exports

pub use bindata::decode_hex_i64_pair;

pub use dates::maybe_date;

Modules

bindata

Decoders for binary data types.

dates

Functions

parse_opt

Parse a possibly-empty string into an option.

trim_opt

Trim a string, and convert to None if it is empty.

trim_owned

Trim a string, and convert to None if it is empty.