Struct bookdata::prelude::MetadataBuilder

pub struct MetadataBuilder<'a> {
    pub(crate) metadata: Metadata<'a>,
}

Builder for Metadata.

Typically should only be used by log library creators or for testing and “shim loggers”. The MetadataBuilder can set the different parameters of a Metadata object, and returns the created object when build is called.

Example

let target = "myApp";
use log::{Level, MetadataBuilder};
let metadata = MetadataBuilder::new()
                    .level(Level::Debug)
                    .target(target)
                    .build();

Fields
metadata: Metadata<'a>
Implementations

impl<'a> MetadataBuilder<'a>

pub fn new() -> MetadataBuilder<'a>

Construct a new MetadataBuilder.

The default options are:

level: Level::Info

target: ""

pub fn level(&mut self, arg: Level) -> &mut MetadataBuilder<'a>

Setter for level.

pub fn target(&mut self, target: &'a str) -> &mut MetadataBuilder<'a>

Setter for target.

pub fn build(&self) -> Metadata<'a>

Returns a Metadata object.