Struct bookdata::prelude::PathBuf

pub struct PathBuf {
    inner: OsString,
}

An owned, mutable path (akin to String).

This type provides methods like push and set_extension that mutate the path in place. It also implements Deref to Path, meaning that all methods on Path slices are available on PathBuf values as well.

More details about the overall approach can be found in the module documentation.

Examples

You can use push to build up a PathBuf from components:

use std::path::PathBuf;

let mut path = PathBuf::new();

path.push(r"C:\");
path.push("windows");
path.push("system32");

path.set_extension("dll");

However, push is best used for dynamic situations. This is a better way to do this when you know all of the components ahead of time:

use std::path::PathBuf;

let path: PathBuf = [r"C:\", "windows", "system32.dll"].iter().collect();

We can still do better than this! Since these are all strings, we can use From::from:

use std::path::PathBuf;

let path = PathBuf::from(r"C:\windows\system32.dll");

Which method works best depends on what kind of situation you’re in.

Fields
inner: OsString
Implementations

impl PathBuf

pub fn new() -> PathBuf

Allocates an empty PathBuf.

Examples

use std::path::PathBuf;

let path = PathBuf::new();

pub fn with_capacity(capacity: usize) -> PathBuf

Creates a new PathBuf with a given capacity used to create the internal OsString. See with_capacity defined on OsString.

Examples

use std::path::PathBuf;

let mut path = PathBuf::with_capacity(10);
let capacity = path.capacity();

// This push is done without reallocating
path.push(r"C:\");

assert_eq!(capacity, path.capacity());

pub fn as_path(&self) -> &Path

Coerces to a Path slice.

Examples

use std::path::{Path, PathBuf};

let p = PathBuf::from("/test");
assert_eq!(Path::new("/test"), p.as_path());

pub fn push<P>(&mut self, path: P)where
P: AsRef<Path>,

Extends self with path.

If path is absolute, it replaces the current path.

On Windows:

if path has a root but no prefix (e.g., \windows), it replaces everything except for the prefix (if any) of self.

if path has a prefix but no root, it replaces self.

if self has a verbatim prefix (e.g. \\?\C:\windows) and path is not empty, the new path is normalized: all references to . and .. are removed.

Examples

Pushing a relative path extends the existing path:

use std::path::PathBuf;

let mut path = PathBuf::from("/tmp");
path.push("file.bk");
assert_eq!(path, PathBuf::from("/tmp/file.bk"));

Pushing an absolute path replaces the existing path:

use std::path::PathBuf;

let mut path = PathBuf::from("/tmp");
path.push("/etc");
assert_eq!(path, PathBuf::from("/etc"));

pub fn pop(&mut self) -> bool

Truncates self to self.parent.

Returns false and does nothing if self.parent is None. Otherwise, returns true.

Examples

use std::path::{Path, PathBuf};

let mut p = PathBuf::from("/spirited/away.rs");

p.pop();
assert_eq!(Path::new("/spirited"), p);
p.pop();
assert_eq!(Path::new("/"), p);

pub fn set_file_name<S>(&mut self, file_name: S)where
S: AsRef<OsStr>,

Updates self.file_name to file_name.

If self.file_name was None, this is equivalent to pushing file_name.

Otherwise it is equivalent to calling pop and then pushing file_name. The new path will be a sibling of the original path. (That is, it will have the same parent.)

Examples

use std::path::PathBuf;

let mut buf = PathBuf::from("/");
assert!(buf.file_name() == None);
buf.set_file_name("bar");
assert!(buf == PathBuf::from("/bar"));
assert!(buf.file_name().is_some());
buf.set_file_name("baz.txt");
assert!(buf == PathBuf::from("/baz.txt"));

pub fn set_extension<S>(&mut self, extension: S) -> boolwhere
S: AsRef<OsStr>,

Updates self.extension to extension.

Returns false and does nothing if self.file_name is None, returns true and updates the extension otherwise.

If self.extension is None, the extension is added; otherwise it is replaced.

Examples

use std::path::{Path, PathBuf};

let mut p = PathBuf::from("/feel/the");

p.set_extension("force");
assert_eq!(Path::new("/feel/the.force"), p.as_path());

p.set_extension("dark_side");
assert_eq!(Path::new("/feel/the.dark_side"), p.as_path());

pub fn as_mut_os_string(&mut self) -> &mut OsString

🔬This is a nightly-only experimental API. (path_as_mut_os_str)

Yields a mutable reference to the underlying OsString instance.

Examples

#![feature(path_as_mut_os_str)]
use std::path::{Path, PathBuf};

let mut path = PathBuf::from("/foo");

path.push("bar");
assert_eq!(path, Path::new("/foo/bar"));

// OsString's `push` does not add a separator.
path.as_mut_os_string().push("baz");
assert_eq!(path, Path::new("/foo/barbaz"));

pub fn into_os_string(self) -> OsString

Consumes the PathBuf, yielding its internal OsString storage.

Examples

use std::path::PathBuf;

let p = PathBuf::from("/the/head");
let os_str = p.into_os_string();

pub fn into_boxed_path(self) -> Box<Path, Global>

Converts this PathBuf into a boxed Path.

pub fn capacity(&self) -> usize

Invokes capacity on the underlying instance of OsString.

pub fn clear(&mut self)

Invokes clear on the underlying instance of OsString.

pub fn reserve(&mut self, additional: usize)

Invokes reserve on the underlying instance of OsString.

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

Invokes try_reserve on the underlying instance of OsString.

pub fn reserve_exact(&mut self, additional: usize)

Invokes reserve_exact on the underlying instance of OsString.

pub fn try_reserve_exact(
&mut self,
additional: usize
) -> Result<(), TryReserveError>

Invokes try_reserve_exact on the underlying instance of OsString.

pub fn shrink_to_fit(&mut self)

Invokes shrink_to_fit on the underlying instance of OsString.

pub fn shrink_to(&mut self, min_capacity: usize)

Invokes shrink_to on the underlying instance of OsString.