Module bookdata::rdf::parse

Parse RDF N-triples.

This module contains a parser for the RDF N-triple format, so we don’t have to pull in all of Oxigraph, can be more robust to bad data formats, and improve parsing performnce. Why do we need PEG to parse n-triples?

Enums

ParseError

Functions

blank_ref

decode_esc

decode_uchars

escape

iri_lit

iri_ref

is_blank_lead

is_blank_mid

is_irichar

is_strchr

lang_tag

literal

nt_line

object

parse_triple

Parse a triple

string_lit

subject

triple

uescape