Module bookdata::util::logging

Logging utilities for the book data tools.

This module contains support for initializing the logging infrastucture, and for dynamically routing log messages based on whether there is an active progress bar.

Structs

Friendly 🔒

Len 🔒

Pos 🔒

Rate 🔒

Constants

DATA_PROGRESS_TMPL 🔒

ITEM_PROGRESS_TMPL 🔒

Traits

FieldExtract 🔒

Functions

data_progress

Create a progress bar for tracking data.

item_progress

Create a progress bar for tracking items.