Module bookdata::cli

Command line interface for book data.

The book data tools are implemented as a single monolithic executable (to reduce compilation time and disk space in common configurations, with different tools implemented as subcommands. Each subcommand implements the Command trait, which exposes the command line arguments and invocation.

Modules

amazon

cluster

Commands for processing clusters.

cluster_books

Book clustering command.

collect_isbns

Collect ISBNs from across the data sources.

extract_graph

Graph extraction program.

fusion

DataFusion script runner.

goodreads

index_names

Index names from authority records.

openlib

scan_marc

Scan MARC records. See ScanMARC for documentation.

Macros

wrap_subcommands

Macro to generate wrappers for subcommand enums.

Structs

AmazonCommandWrapper

CLI

Entry point for the Book Data Tools.

ClusterCommandWrapper

Enums

AmazonCommand

BDCommand

Enum to collect and dispatch CLI commands.

ClusterCommand

Traits

AsyncCommand

Trait for implementing commands asynchronously.

Command

Trait implemented by book data commands.