Module bookdata::cli

Command line interface for book data.

The book data tools are implemented as a single monolithic executable (to reduce compilation time and disk space in common configurations, with different tools implemented as subcommands. Each subcommand implements the Command trait, which exposes the command line arguments and invocation.

Modules

amazon

Amazon commands.

bx

BookCrossing commands.

cluster

Commands for processing clusters.

cluster_books

Book clustering command.

collect_isbns

Collect ISBNs from across the data sources.

extract_graph

Graph extraction program.

filter_marc

Command to filter MARC output.

goodreads

index_names

Index names from authority records.

kcore

link_isbns

openlib

pqinfo

Extract basic information from a Parquet file.

scan_marc

Scan MARC records. See ScanMARC for documentation.

stats

Compute integration statistics.

Macros

wrap_subcommands 🔒

Macro to generate wrappers for subcommand enums.

Structs

AmazonCommandWrapper

BXCommandWrapper

CLI

Entry point for the Book Data Tools.

ClusterCommandWrapper

Enums

AmazonCommand 🔒

BXCommand 🔒

ClusterCommand

RootCommand

Enum to collect and dispatch CLI commands.

Traits

Command

Trait implemented by book data commands.