bookdata::cli::cluster

Module author_gender

Summarize author gender information for clusters.

This script reads the cluster author information and author gender information, in order to aggregate author genders for each cluster.

We use a lot of left joins so that we can compute statistics across the integration pipeline.

Modules

authors 🔒

Support for loading author info.

clusters 🔒

Read cluster information.

Structs

AuthorGender

Extract cluster author gender data from extracted book data.

ClusterGenderInfo 🔒

Record format for saving gender information.

Functions

save_genders 🔒