Function bookdata::cli::filter_marc::scan_records

fn scan_records(
    path: &Path,
    filter: &FilterSpec,
    send: SyncSender<FieldRecord>
) -> Result<JoinHandle<Result<usize>>>

Scan MARC records from a file.

Failes quickly if there is an error opening the file; errors reading the file are from the thread and are availabl when it is joined.