Module bookdata::ids::codes

Book codes and their number spaces.

This implements the book codes for the book integration. In this module, we use book code to refer to the numberspaced numeric code representing the book record in the common code space, and identifier (or underlying identifier) for the native identifier of the underlying record. Identifiers are translated into codes by adding the number space’s base, which partitions the code space into different regions.

Structs

NS

The “number space” structure for identifier spaces.

Constants

NAMESPACES

NS_EDITION

NS_GR_BOOK

NS_GR_WORK

NS_ISBN

NS_LOC_INSTANCE

NS_LOC_REC

NS_LOC_WORK

NS_MULT_BASE

The multiplier base for distinguishing numbers in a number space. Each space supports up to 100K identifiers.

NS_WORK

Functions

add_udfs

Add book code UDFs to an execution context.

ns_of_book_code

Get the namespace for a book code.