bookdata
Module io

Re-exports

pub use compress::open_gzin_progress;

pub use lines::LineProcessor;

pub use object::ObjectWriter;

Modules

background

compress

ext

lines

object

Traits

DataSink

Trait for data processing sinks with input and ouptut files.

Functions

file_size

Get the size of a file.

log_file_info

Log the sizes of a set of files.

open_progress

Open a file as a buffered reader with a progress bar.

path_list

Convert a list of strings into owned PathBufs.