Trait bookdata::io::ext::LengthRead

pub trait LengthRead: Read {
    fn input_size(&self) -> Result<u64>;

    fn read_all_sized(&mut self) -> Result<Vec<u8>> { ... }
}

Trait for read types that can report their size.

Required Methods

fn input_size(&self) -> Result<u64>

Provided Methods

fn read_all_sized(&mut self) -> Result<Vec<u8>>

Read all bytes from a file, using its size to pre-allocate capacity.