Trait bookdata::io::object::ObjectWriter

pub trait ObjectWriter<T>: Sized {
    fn write_object(&mut self, object: T) -> Result<()>;
    fn finish(self) -> Result<usize>;

    fn write_all_objects<I>(&mut self, objects: I) -> Result<usize>
    where
        I: Iterator<Item = T>,
    { ... }
    fn write_and_finish<I>(self, objects: I) -> Result<usize>
    where
        I: Iterator<Item = T>,
    { ... }
    fn with_transform<F, T2>(self, transform: F) -> MapWriter<F, T2, T, Self>
    where
        F: Fn(T2) -> Result<T>,
    { ... }
}

Trait for writing objects to some kind of sink.

Required Methods

fn write_object(&mut self, object: T) -> Result<()>

Write one object.

fn finish(self) -> Result<usize>

Finish and close the target.

Provided Methods

fn write_all_objects<I>(&mut self, objects: I) -> Result<usize>where
I: Iterator<Item = T>,

Write an iterator full of objects.

fn write_and_finish<I>(self, objects: I) -> Result<usize>where
I: Iterator<Item = T>,

Write an iterator of objects and finish the writer.

fn with_transform<F, T2>(self, transform: F) -> MapWriter<F, T2, T, Self>where
F: Fn(T2) -> Result<T>,

Wrap this object writer in a transformed writer.