Trait bookdata::io::DataSink

pub trait DataSink {
    fn output_files(&self) -> Vec<PathBuf>;

    fn input_files(&self) -> Vec<PathBuf> { ... }
}

Trait for data processing sinks with input and ouptut files.

Required methods

fn output_files(&self) -> Vec<PathBuf>

Get the output files for the sink.

Provided methods

fn input_files(&self) -> Vec<PathBuf>

Get auxillary input files for the sink.

Most sinks are also an ObjectWriter, and the primary input is written to the sink; that input file is not reported here. However, sinks may require additional input files to process, and those files can be reported here.