Module bookdata::util::logging

Logging utilities for the book data tools.

This module contains support for initializing the logging infrastucture, and for dynamically routing log messages based on whether there is an active progress bar.

Structs

Friendly 🔒

Len 🔒

Pos 🔒

Rate 🔒

Constants

DATA_PROGRESS_TMPL 🔒

ITEM_PROGRESS_TMPL 🔒

METER_TMPL 🔒

Traits

FieldExtract 🔒

Functions

data_progress

Create a progress bar for tracking data.

item_progress

Create a progress bar for tracking items.

measure_and_recv

Fetch from a receiver while updating the length.

measure_and_send

Send to a channel while updating the length.

meter_bar

Create a meter for monitoring pipelines.